Place prettyprinting code next to syntax declarations
