+++ lang/printer/Cargo.toml
@@ -7,5 +7,3 @@ edition = "2021"
 
 [dependencies]
 pretty = { version = "0.11", features = ["termcolor"] }
-# workspace members
-syntax = { path = "../syntax" }
