+++ lang/printer/src/lib.rs
@@ -6,10 +6,6 @@ pub use pretty::termcolor::StandardStream;
 pub use pretty::termcolor::WriteColor;
 pub use pretty::DocAllocator;
 
-mod ctx;
-mod de_bruijn;
-pub mod fragments;
-mod generic;
 mod render;
 pub mod theme;
 pub mod tokens;
