+++ lang/syntax/src/ast/exp.rs
@@ -4,6 +4,13 @@ use std::rc::Rc;
 
 use codespan::Span;
 use derivative::Derivative;
+use pretty::DocAllocator;
+use printer::theme::ThemeExt;
+use printer::tokens::{
+    ABSURD, ARROW, AS, COLON, COMATCH, COMMA, DOT, FAT_ARROW, HOLE, MATCH, TYPE,
+};
+use printer::util::{BackslashExt, BracesExt, IsNilExt};
+use printer::{Alloc, Builder, Precedence, Print, PrintCfg};
 
 use crate::ctx::values::TypeCtx;
 use crate::ctx::{BindContext, LevelCtx};
@@ -13,6 +20,11 @@ use super::traits::HasSpan;
 use super::traits::Occurs;
 use super::{ident::*, Shift, ShiftRange, ShiftRangeExt};
 use super::{HasType, Leveled};
+
+// Prints "{ }"
+pub fn empty_braces<'a>(alloc: &'a Alloc<'a>) -> Builder<'a> {
+    alloc.space().braces_anno()
+}
 
 #[derive(Debug, Clone, Derivative)]
 #[derivative(Eq, PartialEq, Hash)]
@@ -160,6 +172,40 @@ impl Substitutable for Rc<Exp> {
     }
 }
 
+impl Print for Exp {
+    fn print_prec<'a>(
+        &'a self,
+        cfg: &PrintCfg,
+        alloc: &'a Alloc<'a>,
+        prec: Precedence,
+    ) -> Builder<'a> {
+        match self {
+            Exp::Variable(e) => e.print_prec(cfg, alloc, prec),
+            Exp::TypCtor(e) => e.print_prec(cfg, alloc, prec),
+            Exp::Call(e) => e.print_prec(cfg, alloc, prec),
+            mut dtor @ Exp::DotCall(DotCall { .. }) => {
+                // A series of destructors forms an aligned group
+                let mut dtors_group = alloc.nil();
+                while let Exp::DotCall(DotCall { exp, name, args, .. }) = &dtor {
+                    let psubst = if args.is_empty() { alloc.nil() } else { args.print(cfg, alloc) };
+                    if !dtors_group.is_nil() {
+                        dtors_group = alloc.line_().append(dtors_group);
+                    }
+                    dtors_group =
+                        alloc.text(DOT).append(alloc.dtor(name)).append(psubst).append(dtors_group);
+                    dtor = exp;
+                }
+                dtor.print(cfg, alloc).append(dtors_group.align().group())
+            }
+            Exp::Anno(e) => e.print_prec(cfg, alloc, prec),
+            Exp::TypeUniv(e) => e.print_prec(cfg, alloc, prec),
+            Exp::LocalMatch(e) => e.print_prec(cfg, alloc, prec),
+            Exp::LocalComatch(e) => e.print_prec(cfg, alloc, prec),
+            Exp::Hole(e) => e.print_prec(cfg, alloc, prec),
+        }
+    }
+}
+
 // Variable
 //
 //
@@ -239,6 +285,24 @@ impl Substitutable for Variable {
     }
 }
 
+impl Print for Variable {
+    fn print_prec<'a>(
+        &'a self,
+        cfg: &PrintCfg,
+        alloc: &'a Alloc<'a>,
+        _prec: Precedence,
+    ) -> Builder<'a> {
+        let Variable { name, idx, .. } = self;
+        if cfg.de_bruijn {
+            alloc.text(format!("{name}@{idx}"))
+        } else if name.is_empty() {
+            alloc.text(format!("@{idx}"))
+        } else {
+            alloc.text(name)
+        }
+    }
+}
+
 // TypCtor
 //
 //
@@ -306,6 +370,30 @@ impl Substitutable for TypCtor {
     fn subst<S: Substitution>(&self, ctx: &mut LevelCtx, by: &S) -> Self {
         let TypCtor { span, name, args } = self;
         TypCtor { span: *span, name: name.clone(), args: args.subst(ctx, by) }
+    }
+}
+
+impl Print for TypCtor {
+    fn print_prec<'a>(
+        &'a self,
+        cfg: &PrintCfg,
+        alloc: &'a Alloc<'a>,
+        prec: Precedence,
+    ) -> Builder<'a> {
+        let TypCtor { span: _, name, args } = self;
+        if name == "Fun" && args.len() == 2 && cfg.print_function_sugar {
+            let arg = args.args[0].print_prec(cfg, alloc, 1);
+            let res = args.args[1].print_prec(cfg, alloc, 0);
+            let fun = arg.append(alloc.space()).append(ARROW).append(alloc.space()).append(res);
+            if prec == 0 {
+                fun
+            } else {
+                fun.parens()
+            }
+        } else {
+            let psubst = if args.is_empty() { alloc.nil() } else { args.print(cfg, alloc) };
+            alloc.typ(name).append(psubst)
+        }
     }
 }
 
@@ -396,6 +484,19 @@ impl Substitutable for Call {
             args: args.subst(ctx, by),
             inferred_type: None,
         }
+    }
+}
+
+impl Print for Call {
+    fn print_prec<'a>(
+        &'a self,
+        cfg: &PrintCfg,
+        alloc: &'a Alloc<'a>,
+        _prec: Precedence,
+    ) -> Builder<'a> {
+        let Call { name, args, .. } = self;
+        let psubst = if args.is_empty() { alloc.nil() } else { args.print(cfg, alloc) };
+        alloc.ctor(name).append(psubst)
     }
 }
 
@@ -565,6 +666,18 @@ impl Substitutable for Anno {
     }
 }
 
+impl Print for Anno {
+    fn print_prec<'a>(
+        &'a self,
+        cfg: &PrintCfg,
+        alloc: &'a Alloc<'a>,
+        _prec: Precedence,
+    ) -> Builder<'a> {
+        let Anno { exp, typ, .. } = self;
+        exp.print(cfg, alloc).parens().append(COLON).append(typ.print(cfg, alloc))
+    }
+}
+
 // TypeUniv
 //
 //
@@ -630,6 +743,17 @@ impl Substitutable for TypeUniv {
     fn subst<S: Substitution>(&self, _ctx: &mut LevelCtx, _by: &S) -> Self::Result {
         let TypeUniv { span } = self;
         TypeUniv { span: *span }
+    }
+}
+
+impl Print for TypeUniv {
+    fn print_prec<'a>(
+        &'a self,
+        _cfg: &PrintCfg,
+        alloc: &'a Alloc<'a>,
+        _prec: Precedence,
+    ) -> Builder<'a> {
+        alloc.keyword(TYPE)
     }
 }
 
@@ -712,6 +836,28 @@ impl Substitutable for LocalMatch {
     }
 }
 
+impl Print for LocalMatch {
+    fn print_prec<'a>(
+        &'a self,
+        cfg: &PrintCfg,
+        alloc: &'a Alloc<'a>,
+        _prec: Precedence,
+    ) -> Builder<'a> {
+        let LocalMatch { name, on_exp, motive, cases, .. } = self;
+        on_exp
+            .print(cfg, alloc)
+            .append(DOT)
+            .append(alloc.keyword(MATCH))
+            .append(match &name.user_name {
+                Some(name) => alloc.space().append(alloc.dtor(name)),
+                None => alloc.nil(),
+            })
+            .append(motive.as_ref().map(|m| m.print(cfg, alloc)).unwrap_or(alloc.nil()))
+            .append(alloc.space())
+            .append(print_cases(cases, cfg, alloc))
+    }
+}
+
 // LocalComatch
 //
 //
@@ -783,6 +929,48 @@ impl Substitutable for LocalComatch {
             inferred_type: None,
         }
     }
+}
+
+impl Print for LocalComatch {
+    fn print_prec<'a>(
+        &'a self,
+        cfg: &PrintCfg,
+        alloc: &'a Alloc<'a>,
+        _prec: Precedence,
+    ) -> Builder<'a> {
+        let LocalComatch { name, is_lambda_sugar, cases, .. } = self;
+        if *is_lambda_sugar && cfg.print_lambda_sugar {
+            print_lambda_sugar(cases, cfg, alloc)
+        } else {
+            alloc
+                .keyword(COMATCH)
+                .append(match &name.user_name {
+                    Some(name) => alloc.space().append(alloc.ctor(name)),
+                    None => alloc.nil(),
+                })
+                .append(alloc.space())
+                .append(print_cases(cases, cfg, alloc))
+        }
+    }
+}
+
+/// Print the Comatch as a lambda abstraction.
+/// Only invoke this function if the comatch contains exactly
+/// one cocase "ap" with three arguments; the function will
+/// panic otherwise.
+fn print_lambda_sugar<'a>(cases: &'a [Case], cfg: &PrintCfg, alloc: &'a Alloc<'a>) -> Builder<'a> {
+    let Case { params, body, .. } = cases.first().expect("Empty comatch marked as lambda sugar");
+    let var_name = params
+        .params
+        .get(2) // The variable we want to print is at the third position: comatch { ap(_,_,x) => ...}
+        .expect("No parameter bound in comatch marked as lambda sugar")
+        .name();
+    alloc
+        .backslash_anno(cfg)
+        .append(var_name)
+        .append(DOT)
+        .append(alloc.space())
+        .append(body.print(cfg, alloc))
 }
 
 // Hole
@@ -869,6 +1057,21 @@ impl Substitutable for Hole {
     }
 }
 
+impl Print for Hole {
+    fn print_prec<'a>(
+        &'a self,
+        cfg: &PrintCfg,
+        alloc: &'a Alloc<'a>,
+        _prec: Precedence,
+    ) -> Builder<'a> {
+        if cfg.print_metavar_ids {
+            alloc.text(format!("?{}", self.metavar.id))
+        } else {
+            alloc.keyword(HOLE)
+        }
+    }
+}
+
 // Case
 //
 //
@@ -916,6 +1119,46 @@ impl Substitutable for Case {
     }
 }
 
+impl Print for Case {
+    fn print<'a>(&'a self, cfg: &PrintCfg, alloc: &'a Alloc<'a>) -> Builder<'a> {
+        let Case { span: _, name, params, body } = self;
+
+        let body = match body {
+            None => alloc.keyword(ABSURD),
+            Some(body) => alloc
+                .text(FAT_ARROW)
+                .append(alloc.line())
+                .append(body.print(cfg, alloc))
+                .nest(cfg.indent),
+        };
+
+        alloc.ctor(name).append(params.print(cfg, alloc)).append(alloc.space()).append(body).group()
+    }
+}
+
+pub fn print_cases<'a>(cases: &'a [Case], cfg: &PrintCfg, alloc: &'a Alloc<'a>) -> Builder<'a> {
+    match cases.len() {
+        0 => empty_braces(alloc),
+
+        1 => alloc
+            .line()
+            .append(cases[0].print(cfg, alloc))
+            .nest(cfg.indent)
+            .append(alloc.line())
+            .braces_anno()
+            .group(),
+        _ => {
+            let sep = alloc.text(COMMA).append(alloc.hardline());
+            alloc
+                .hardline()
+                .append(alloc.intersperse(cases.iter().map(|x| x.print(cfg, alloc)), sep.clone()))
+                .nest(cfg.indent)
+                .append(alloc.hardline())
+                .braces_anno()
+        }
+    }
+}
+
 // Telescope Inst
 //
 //
@@ -934,6 +1177,16 @@ impl TelescopeInst {
 
     pub fn is_empty(&self) -> bool {
         self.params.is_empty()
+    }
+}
+
+impl Print for TelescopeInst {
+    fn print<'a>(&'a self, cfg: &PrintCfg, alloc: &'a Alloc<'a>) -> Builder<'a> {
+        if self.params.is_empty() {
+            alloc.nil()
+        } else {
+            self.params.print(cfg, alloc).parens()
+        }
     }
 }
 
@@ -958,6 +1211,13 @@ pub struct ParamInst {
 impl Named for ParamInst {
     fn name(&self) -> &Ident {
         &self.name
+    }
+}
+
+impl Print for ParamInst {
+    fn print<'a>(&'a self, _cfg: &PrintCfg, alloc: &'a Alloc<'a>) -> Builder<'a> {
+        let ParamInst { span: _, info: _, name, typ: _ } = self;
+        alloc.text(name)
     }
 }
 
@@ -1000,6 +1260,20 @@ impl Substitutable for Args {
     }
 }
 
+impl Print for Args {
+    fn print<'a>(&'a self, cfg: &PrintCfg, alloc: &'a Alloc<'a>) -> Builder<'a> {
+        let mut doc = alloc.nil();
+        let mut iter = self.args.iter().peekable();
+        while let Some(arg) = iter.next() {
+            doc = doc.append(arg.print(cfg, alloc));
+            if iter.peek().is_some() {
+                doc = doc.append(COMMA).append(alloc.line())
+            }
+        }
+        doc.align().parens().group()
+    }
+}
+
 // Motive
 //
 //
@@ -1035,5 +1309,21 @@ impl Substitutable for Motive {
             param: param.clone(),
             ret_typ: ctx.bind_single((), |ctx| ret_typ.subst(ctx, &by.shift((1, 0)))),
         }
+    }
+}
+
+impl Print for Motive {
+    fn print<'a>(&'a self, cfg: &PrintCfg, alloc: &'a Alloc<'a>) -> Builder<'a> {
+        let Motive { span: _, param, ret_typ } = self;
+
+        alloc
+            .space()
+            .append(alloc.keyword(AS))
+            .append(alloc.space())
+            .append(param.print(cfg, alloc))
+            .append(alloc.space())
+            .append(alloc.text(FAT_ARROW))
+            .append(alloc.space())
+            .append(ret_typ.print(cfg, alloc))
     }
 }
